+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Add new flavour with workflows and makefile for cluster apps.
+
+### Changed
+
 - repo setup: better select checks required for PR merge
 
 ## [5.19.0] - 2023-02-21